university-chatbot/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
python-docx
